Provide a more dynamic Actions management solution
Today, the set of possible `Actions`_ is fixed for a given version of Watcher
and enables optimization algorithms to include actions such as instance
migration, changing hypervisor state, changing power state (ACPI level, ...).

It should be possible to deploy and configure (in watcher.conf) new potential
`Actions`_ and associate each `Action`_ to a Python class, without the need to
redeploy a new version of Watcher.

Change-Id: If4b68079967854f9040f11206e2b6202a66d8551
blueprint: watcher-add-actions-via-conf