Define the scope of an Audit as pool of resources
As a cloud administrator, I want to be able to define the pool of
resources on which an Audit will apply.
This scope will be called "Audit scope" and will be configured in an
Audit Template as a list of managed resource ids.

It will enable Watcher to provide a very generic auditing system which
will be able to audit:
- any kind of resources: compute nodes, host aggregates, storage nodes,
  network equipments, network links, Hadoop cluster, ...
- for any kinf of goal (energy optimization, servers consolidation,
  network latency optimization, network bandwidth optimization, ...).

Change-Id: I8fe30a187c1938c1cc35ecf213603b236ac5fc8f
blueprint: define-the-audit-scope