Provide efficacy indicators
When the admin trigger an audit using Watcher, we provide a list of
actions that should be run to attain a specific goal (defined in the
audit template), Today, there is no incentive to give a predicted
result of the optimization after running the actions plan. This
feature will give the admin an estimated gain of running Watcher
on the infrastructure.

Change-Id: I774b6bde9329135d7cb9c85387ea40f9ab59d49f
Implements: blueprint bp/efficacy-indicator