Cluster model objects wrapper
Currently a model of the cluster is constructed every time a strategy is
executed, which will not scale in production or in larger environments.
If Watcher intends to be used by larger envirionments it needs a more
robust way to construct and maintain a model of the cluster. An
in-memory cache of this model can be built up and kept fresh via
notifications from services of interest in addition to periodic syncing
logic.

Co-Authored-By: Vincent FRANCOISE <Vincent.FRANCOISE@b-com.com>
Change-Id: I338b1169049c80638c5229648c6f26ed39a9b622