Achieved goal should be returned by each strategy
Today, there is no way to know what goal is achieved by a given
strategy. The mapping between goal and strategies is done in the
Watcher configuration.

Beyond that, there is no way to know that several strategies achieve
the same optimization goal.
There should be a way to return the list of available goals depending
on which strategies have been deployed on the node where the Watcher
Decision Engine is running.

Change-Id: I57f7e434c41fcbd6606552c186e9a0285ff26e97
Implements: blueprint get-goal-from-strategy