Audit parameters should be persistent in Db
When you start an audit, a new Audit object
is created into Db with a reference to an Audit
Template object.
If we update the Audit Template parameters or
if we remove the Audit Template, we lose
important information about the Audit itself.
What would be the consequence for a
periodic audit ? What about data integrity for
soft-deleted audit objects (useful for history)?
If we introduce parameters for goals
(as thresholds, ...) , we should be able to
update them. So how to keep persistent data
for all audits done to satisfy such
configurable goals ?

Change-Id: Ia3560bba42f3d9b9cbccc6795cc41ce887a4c0bb