Define grammar for workload characterization
As we run several workloads in cloud, we should be able to characterize such
workloads as input to watcher for ensuring Application QoS, placements and
consolidation.

An example of workload characterization is a weighted combination of CPU,
Memory or any other resource attributes like High IOPs, Network latency etc.

Scope of this blueprint is to come up with a grammar structure for defining
workload character.

Co-Authored-By: Chris Spencer <christopher.m.spencer@intel.com>
Co-Authored-By: Susanne Balle <sleipnir012@gmail.com>
Co-Authored-By: Prashanth Hari <hvprash@gmail.com>
blueprint: workload-characterization-grammar
Change-Id: I6f43b49042371fc2feac17db2be5861a89fe5262