Define when an action plan is stale/invalid
When an audit is created and launched successfully, it generates
a new Action Plan with status RECOMMENDED. If the Cluster Data
Model has changed by and by, the action plan is still keep the
RECOMMENDED state. There is not an expiry date or event that can
invalidate the action plan by far.
It should be possible to add a new "SUPERSEDED" state for Action
Plan and Action.

Change-Id: I875fbcb44876114b69fc1eaf4729aa7b292b399c
Implements: blueprint stale-action-plan