Audit tag in VM Metadata
When watcher runs audits to achieve a goal, there should be some way for the
application/VM owners to know that their VMs are under audit and is flagged
for action plan execution. We could store these information in VM metadata
with a timestamp after which action plan will be executed.

blueprint: audit-tag-vm-metadata
Change-Id: I5ef4ce84057acee46f09bed684608f303ccd8291