Support Watcher HA active-active Mode
Watcher-decision-engine and Applier currently don't support
HA active-active mode.
If we enable more than one Watcher-decision-engine, the question is
how to know the relation between audit and Watcher-decision-engine.
We can add a new 'host' field in the audit table to solve
this question. When we update the audit sate from PENDING
to ONGOING, we record the host into the 'host' field.

Change-Id: I355e478bb0171a67bde8b95468e193e274c1fd0c
Implements: blueprint support-watcher-ha-active-active-mode