General purpose threadpool for the decision engine
Spec to create a general purpose threadpool to be implemented in the
decision engine. The amount of threads is configurable to best fit
different OpenStack deployments while the default value should be
conservative to prevent any issues.

Additionally, the building of the data model will use the threadpool
and will demonstrate its usefulness.

Test cases will be made for both the threadpool as well as for the
behavior of building the data model.

Change-Id: Iaaaea3bc436d405cf19f99b13ed4804e10e646c6