Add status_message field to the Audits
As part of the spec to add a mechanism to skip actions we are adding a
new field `status_message` to actions and actionplans. In the
discussions about it, we mentioned to also add it to the audits and
cover it in the same API microversion but it was not covered by the
spec.

This patch is adding the `status_message` field in the Audits to the
spec document.

blueprint add-skip-actions

Change-Id: Ifd663042386304a19962710c182644f7fdfe864a
Signed-off-by: Alfredo Moralejo <amoralej@redhat.com>