Change Aetos's service type
We were told, we need to register any new service types in
service-types-authority and we saw, there is a naming convention
for new types. We decided to rename the service type used for
Aetos from 'prometheus' to 'metric-storage' and register it.
As part of this, we're dropping the requirement for the service
name to be 'aetos', which is a bit more in line with other
services.

The spec is updated to reflect these changes

Depends-On: https://review.opendev.org/c/openstack/service-types-authority/+/955782

Change-Id: I75cc5f8961983eca2c2521213312cb850c004f68
Signed-off-by: Jaromir Wysoglad <jwysogla@redhat.com>